增加小铺用户提现和账户操作记录的service代码

+++ shop-dao/src/main/java/com/masiis/shop/dao/po/SfUserExtractApply.java
@@ -32,7 +32,7 @@ public class SfUserExtractApply {
     /**
      * 1,微信; 2,支付宝; 3,银行卡
      */
-    private Byte extractWay;
+    private Integer extractWay;
     /**
      * 审核通过时间
      */
@@ -97,10 +97,10 @@ public class SfUserExtractApply {
     public void setApplyTime(Date applyTime) {
         this.applyTime = applyTime;
     }
-    public Byte getExtractWay() {
+    public Integer getExtractWay() {
         return extractWay;
     }
-    public void setExtractWay(Byte extractWay) {
+    public void setExtractWay(Integer extractWay) {
         this.extractWay = extractWay;
     }
     public Date getAuditTime() {
